LayoutTests/media/media-fragments/TC0001-TC0009-expected.txt


Title: TC0001
Fragment: 't=,'
Comment: Syntax error, not allowed according to the ABNF. The media fragment is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK

Title: TC0002
Fragment: 't=3,3'
Comment: Invalid semantics: start must be smaller than end. The media fragment is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK

Title: TC0003
Fragment: 't=7,3'
Comment: Invalid semantics: the requested interval's start is beyond its end. The media fragment is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK

Title: TC0004
Fragment: 't=0,9.97'
Comment: The media is requested from 0 to e.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0005
Fragment: 't=3,7'
Comment: The media is requested from a to b.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0006
Fragment: 't=3,15'
Comment: The media is requested from a to e.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0009
Fragment: 't=15,20'
Comment: The request lies beyond the end of the resource. If the UA knows the duration of the resource, it seeks to the end of the media resource. Otherwise, the UA will send an (out-of-range) HTTP request with an 'include-setup' in order to setup its decoding pipeline.
EVENT(canplaythrough)
EXPECTED (video.currentTime == 'duration') OK
EXPECTED (video.initialTime == 'duration') OK

END OF TEST
